lint.sh: Allow to lint-ignore through Jinja comments
Previously, we required a line that should be ignored by the linter to
end in `lint-ignore`. This is problematic in lines that cannot have a
proper comment character. Hence, we allow to lint-ignore through

  {# lint-ignore #}

at the end of a line. This helps for example with the `debug_print` in
the Exim config of the `smtp-server` role.