apache-website: Move apache default website into dedicated role
Without the default website being its own role, we run into issues
when setting up another website. Say we want to set up a grafana site
through apache-website. This would depend on apache-webserver, which
would inturn depend on apache_website to set up the default
site. Since variables in Ansible are global, the variables from the
default site would override the variables given for the Grafana
website.

To avoid this overriding, we turn the default site into a dedicated
role. This avoids apache-website indirectly depending on itself and
hence avoids the overriding of variables.