Switch to installing Apache webserver late
If we install Apache webserver unconditionally early, we might install
it needlessly (as we cannot reliably detect upfront if a host will
need an apache webserver or not).

Including it conditionally in website roles through meta no longer
works reliably in recent Ansible versions and did not get run in some
settings when running only for a single tag. Working around this
through global reference counting fails too with recent Ansibles when
using more than a single host, as host name is not yet available when
filters run and recursion detection does not take conditionals into
account.

So we resort to installing apache late. This will reliably install the
Apache server only when needed, but requires to pass around another
list of domains, as we'd otherwise could not detect which domains got
selected by running only for some tags.