Rename `host_info_os` to `qhost_distribution`
It is fed by `ansible_distribution`, so having it end in
`_distribution` instead of `_os` better matches Ansibles conventions.

`host_info` seemed inconspicious. So we have it start with a `q`,
which sticks out more. And it helps with grepping, as `host_info`
sounds more like a variable someone else might want to use at some
point.