icinga-server: Alert later about expiring certificates
The previous 70/56 days were fine if most certificates are valid for a
year and changes needed coordination with clients. But since the world
switched to Let's Encrypt, certs are typically valid only for 90 days
and are getting switched often. Hence, client coordination is
typically no longer needed, as clients adjusted to ever changing
certs. So we shorten the alerting periods to 14/2 days.