apache-website: Add clarification for grant within RequireAll
For a `RequireAll` to succeed, at least one of its children needs to
succeed. It we would not have this `Require all granted`, it might be
that the `RequireAll` does not have children at all, and would hence
fail. By adding the `Require all granted` we make sure the
`RequireAll` has children and hence passes if there are not other
children.