common-role-tasks: Move setting ownership to start
Now that the start role creates the necessary user and group, we can
bring the ownership setting to the start role as well, as this allows
to bring files to the server under the correct user and group. This is
necessary to somewhat confidentially bring config files that contain
passwords to the host.