common-role-tasks: Omit defaults for most options
While most of the configuration options would allow for good defaults,
providing such defaults invites for overlooking settings. E.g.:
Providing a default for port configs would invite to overlook the need
to set this variable when writing a new role that relies of
networking. Requests would get eaten in the firewall, and it would not
necessarily be apparent to users what is going on.

So we don't provide such defaults. Then users that write new roles are
forced to thing about options like port configs and can quickly set
them to empty values if unneeded.