diamond: Add user to adm group unconditionally.
We need adm not only for apache, and nginx, but also for access to
openvpn logs and many more log files. Updating the required condition
for `adm` membership became a burden. As effectively each host has
diamond in the adm group for one reason or the other, we simply add it
to the adm group unconditionally. This is an effective no-op on hosts,
but reduced maintenance burden a lot.