apache-website: Use effective instead of logical apache log dir
When websites get set up, the apache server did not yet get set up. So
/var/log/apache2 is not yet the proper symlink that it'll be in the
end. So instead of creating the the symlink in all possible flavors of
apache sites, we instead set up the site log dir directly in the
effective log dir. That way, we don't need to mess with the link
creation in /var/log.