Monitoring: Move from `variant` to `max_seconds`
For some sites, the 3 seconds of the `slow` variants were too
short. So we add more variants with longer timeouts. To get proper
alerting, we raise from warnings to errors. Also, since this setting
is seldomly used, it's hard to remember which mnemonic identifier
stands for which value, so we switch to the number of seconds, as that
is more descriptive.