graphite: Store logs of webapp and carbon-cache in separate directories
Since both the webapp and carbon-cache belong to graphite, we had logs
of both applications in one directory. But as both applications use
rather generic names for their log files (E.g.:
info.log. exception.log), this made it hard to see which file is from
which application. Hence, we switch to one directory per application.