common-packages: Add acl
`icinga-server` forces owner and group for log files, while we need to
grant access to `nagios`, `www-data`, and `adm`. Since we don' want to
make the world readable, using acls is the cleanest solution. But
Ansible's `acl` package relies on the `getfacl` and `setfacl`
binaries, which are not installed by default on Debian. Hence, we
install the `acl` package, which provides the needed binaries.